Update to current clang-format
I had been working in the clang repository, but do to the way patch
creation works, I think it's cleaner to just workin a repository with
the file I'm editing.

Since the other repo got so mangled, I'm just bringing in the most
up-to-date version, rather than try to recreate the commits.